refactor: split store.rs into focused modules (#179)
Extract store.rs (2,065 LOC) into 6 focused modules:
- schema.rs (183 LOC) — init, migration, versioning
- crud.rs (234 LOC) — insert, get, delete, update, list, search, count
- tags.rs (184 LOC) — unique_tags, tags_with_counts, rename_tag, delete_tag
- aging.rs (147 LOC) — touch_access, find_aged, cleanup_aged, tier_counts
- bulk.rs (150 LOC) — bulk_delete_by_tag, bulk_delete_cold, bulk_delete_all, deprecate, prune_ttl

store.rs slimmed to 985 LOC (struct, open(), helpers, all tests).
Zero behavioral changes — pure reorganization.

Note: Cannot verify build locally (missing pkg-config for ort/openssl-sys).
CI will validate compilation and test pass.
