+++ crates/uteke-core/src/memory/mod.rs
@@ -1,6 +1,11 @@
 //! Memory storage and retrieval components.
 
+pub mod aging;
+pub mod bulk;
+pub mod crud;
+pub mod schema;
 pub mod store;
+pub mod tags;
 pub mod types;
 pub mod vector;
